feat(listening): 添加获取听力练习功能

+++ src/main/java/com/code/cetboot/dto/ReadingPracticeDTO.java
@@ -16,5 +16,5 @@ public class ReadingPracticeDTO {
     /**
      * 阅读练习id
      */
-    private long readingPracticeId;
+    private Integer readingPracticeId;
 }
